+++ src/test/GEOSTest/Sample.hs
@@ -3,8 +3,6 @@
 module GEOSTest.Sample (demo) where
 
 import Control.Monad
-import Control.Monad.Trans
-import Control.Monad.Trans.Maybe
 import Data.Geolocation.GEOS
 import Data.List
 import Data.Maybe
@@ -27,13 +25,13 @@ data Extent = Extent
     , maxZ :: Double
     } deriving Show
 
-getXYZs :: CoordinateSequence -> MaybeT IO [Coordinate]
+getXYZs :: CoordinateSequence -> IO [Coordinate]
 getXYZs coordSeq = do
-    size <- MaybeT $ getSize coordSeq
+    size <- getSize coordSeq
     forM [0..(size - 1)] $ \i -> do
-        x <- MaybeT $ getX coordSeq i
-        y <- MaybeT $ getY coordSeq i
-        z <- MaybeT $ getZ coordSeq i
+        x <- getX coordSeq i
+        y <- getY coordSeq i
+        z <- getZ coordSeq i
         return (x, y, z)
 
 extent :: [Coordinate] -> Extent
@@ -49,7 +47,7 @@ extent ((x, y, z) : cs) =
     in Extent minX maxX minY maxY minZ maxZ
 
 -- TODO: Build polygon using API instead of string parsing
-mkSquare :: Reader -> Longitude -> Latitude -> Resolution -> MaybeT IO Geometry
+mkSquare :: Reader -> Longitude -> Latitude -> Resolution -> IO Geometry
 mkSquare reader longitude latitude resolution =
     let points = [
             (longitude, latitude),
@@ -59,32 +57,32 @@ mkSquare reader longitude latitude resolution =
             (longitude, latitude)
             ]
         wkt = printf "POLYGON ((%s))" (intercalate "," (map (\(a, b) -> printf "%f %f" a b) points))
-    in MaybeT $ readGeometry reader wkt
+    in readGeometry reader wkt
 
-getGeometries :: Geometry -> IO (Maybe [Geometry])
-getGeometries geometry = runMaybeT $ do
-    count <- MaybeT $ getNumGeometries geometry
-    mapM (MaybeT . getGeometry geometry) [0..(count - 1)]
+getGeometries :: Geometry -> IO [Geometry]
+getGeometries geometry = do
+    count <- getNumGeometries geometry
+    mapM (getGeometry geometry) [0..(count - 1)]
 
-findBiggestPolygon :: Geometry -> IO (Maybe Geometry)
-findBiggestPolygon geometry = runMaybeT $ do
-    gs@(gHead : gTail) <- MaybeT (getGeometries geometry)
-    as@(aHead : aTail) <- mapM (MaybeT . area) gs
+findBiggestPolygon :: Geometry -> IO Geometry
+findBiggestPolygon geometry = do
+    gs@(gHead : gTail) <- getGeometries geometry
+    as@(aHead : aTail) <- mapM area gs
     return $ fst $ foldr
         (\p'@(_, a') p@(_, a) -> if a' > a then p' else p)
         (gHead, aHead)
         (zip gTail aTail)
 
-processPolygon :: String -> Resolution -> Longitude -> Latitude -> Geometry -> MaybeT IO ()
+processPolygon :: String -> Resolution -> Longitude -> Latitude -> Geometry -> IO ()
 processPolygon tableName resolution longitude latitude p = do
-    shell <- MaybeT $ getExteriorRing p
-    coordSeq <- MaybeT $ getCoordSeq shell
+    shell <- getExteriorRing p
+    coordSeq <- getCoordSeq shell
     xyzs <- getXYZs coordSeq
     let pId = polygonId resolution longitude latitude
         format :: Int -> Double -> Double -> String
         format = printf "INSERT INTO %s (polygon_id, point_id, longitude, latitude) VALUES ('%s', %d, %f, %f);\n" tableName pId
         strs = map (\(pointId, (x, y, _)) -> format pointId x y) (zip [0..] xyzs)
-    lift $ forM_ strs putStr
+    forM_ strs putStr
 
 polygonId :: Resolution -> Longitude -> Latitude -> String
 polygonId resolution longitude latitude = "id_" ++ formatValue longitude ++ "_" ++ formatValue latitude
@@ -92,19 +90,18 @@ polygonId resolution longitude latitude = "id_" ++ formatValue longitude ++ "_" 
         formatValue :: Double -> String
         formatValue = replace "-" "m" . replace "." "_" . (printf "%.2f") . mfloor resolution
 
-generatePolygonMeshSQL :: Context -> String -> Resolution -> MaybeT IO ()
+generatePolygonMeshSQL :: Context -> String -> Resolution -> IO ()
 generatePolygonMeshSQL ctx wkt resolution = do
-    reader <- MaybeT $ mkReader ctx
-    writer <- MaybeT $ mkWriter ctx
-    country <- MaybeT $ readGeometry reader wkt
-    env <- MaybeT $ envelope country
-    shell <- MaybeT $ getExteriorRing env
-    coordSeq <- MaybeT $ getCoordSeq shell
+    reader <- mkReader ctx
+    writer <- mkWriter ctx
+    country <- readGeometry reader wkt
+    env <- envelope country
+    shell <- getExteriorRing env
+    coordSeq <- getCoordSeq shell
     xyzs <- getXYZs coordSeq
 
-    lift $ do
-        putStrLn "Shell:"
-        mapM_ print xyzs
+    putStrLn "Shell:"
+    mapM_ print xyzs
 
     let Extent{..} = extent xyzs
         mfloorRes = mfloor resolution
@@ -113,21 +110,20 @@ generatePolygonMeshSQL ctx wkt resolution = do
         latitudeBegin = mfloorRes minY
         latitudeEnd = mfloorRes maxY + resolution
 
-    lift $ do
-        putStrLn "Longitude and latitude ranges:"
-        putStrLn $ printf "  longitude %f to %f" longitudeBegin longitudeEnd
-        putStrLn $ printf "  latitude %f to %f" latitudeBegin latitudeEnd
+    putStrLn "Longitude and latitude ranges:"
+    putStrLn $ printf "  longitude %f to %f" longitudeBegin longitudeEnd
+    putStrLn $ printf "  latitude %f to %f" latitudeBegin latitudeEnd
 
     let longitudes = frange longitudeBegin longitudeEnd 1.0
         latitudes = frange latitudeBegin latitudeEnd 1.0
     forM_ [(i, j) | i <- longitudes, j <- latitudes] $ \(longitude, latitude) -> do
         square <- mkSquare reader longitude latitude resolution
-        overlap <- MaybeT $ intersection square country
-        isOverlapEmpty <- MaybeT $ isEmpty overlap
+        overlap <- intersection square country
+        isOverlapEmpty <- isEmpty overlap
         unless isOverlapEmpty $ do
-            typeId <- MaybeT $ geomTypeId overlap
+            typeId <- geomTypeId overlap
             polygon <- case typeId of
-                            MultiPolygon -> MaybeT $ findBiggestPolygon overlap
+                            MultiPolygon -> findBiggestPolygon overlap
                             Polygon -> return overlap
             processPolygon "TABLE_NAME" resolution longitude latitude polygon
     return ()
@@ -136,6 +132,5 @@ demo :: IO ()
 demo = do
     fileName <- getDataFileName "data/namibia.wkt"
     wkt <- readFile fileName
-    result <- withGEOS $ \ctx -> runMaybeT (generatePolygonMeshSQL ctx wkt 1.0)
-
-    putStrLn $ "Sample.demo: " ++ (if isJust result then "succeeded" else error "Sample.demo failed")
+    withGEOS $ \ctx -> generatePolygonMeshSQL ctx wkt 1.0
+    putStrLn "Sample.demo: succeeded"