+++ src/test/Main.hs
@@ -3,16 +3,16 @@
 module Main (main) where
 
 import Data.Geolocation.GEOS
-import qualified GEOSTest.HighLevelAPI as HighLevelAPI
+import qualified GEOSTest.HighLevelAPI1 as HighLevelAPI1
+import qualified GEOSTest.HighLevelAPI2 as HighLevelAPI2
 import qualified GEOSTest.LowLevelAPI as LowLevelAPI
 import qualified GEOSTest.Sample as Sample
-import qualified GEOSTest.TransAPI as TransAPI
 
 main :: IO ()
 main = do
     v <- version
     putStrLn v
     LowLevelAPI.demo
-    HighLevelAPI.demo
-    TransAPI.demo
+    HighLevelAPI1.demo
+    HighLevelAPI2.demo
     Sample.demo